Rename GameServer to ServerMemory, to reflect the desired intention of the class: the memory of the server
